requirements.txt
anthropic
langgraph
langchain
langchain-anthropic
langchain-community
tavily-python
python-dotenv
pymupdf
streamlit
pandas
pytest
datasets
scikit-learn
torch
numpy<2
sentence-transformers==2.7.0
python-docx
